Add psuedocode and update readme for exploring Auth0 Management API
This commit captures examples of some of the entities that I have
commonly needed to create. I added the repl namespace to play with
hitting the Auth0 API, and captured the results in the example edn
file. These aren't actually being used yet, but I outlined the goal
in the readme so that once it is working, the main ideas will already
be stated.