Move build-url, get-entities, and get-entity to auth0 ns
The move was done since it was providing a minimal code/ maximal reuse
value in the repl namespace. The repl code was altered to use more
convenient versions of the function (not having to provide domain and
token if you just want to quickly be able to investigate an existing
Auth0 environment).