Add snapshot function to allow current config to be pulled from Auth0
Updated the README to demonstrate how to use the snapshot function,
and updated the main program in order to allow it to be run with
configuration either from the environment variables, or to allow the
user to call it as a function to provide clojure data. This is in
anticipation of using this to support use as a release tool.