Revamp how program output is generated, detect Auth0 api errors
There is still no attempt to rollback, which would require more
effort.

The format-output function now allows the user to chose the type of
output that they want, HR (Human Readable) and EDN. The get-edn-output
multimethod was added to make creating the output easier to manage.

Couple other minor fixes. One to http-helper to have more of the
options it needs to perform a get. Another to remove a pprint of opts
that was mistakenly left in.